+++ src/store/socket.js
@@ -699,7 +699,8 @@ class LiveSession {
     const players = this._store.state.players.players;
     if (
       !nomination ||
-      (players.length > nomination[0] && players.length > nomination[1])
+      ((typeof nomination[0] !== "number" || players.length > nomination[0]) &&
+        (typeof nomination[1] !== "number" || players.length > nomination[1]))
     ) {
       this.setVotingSpeed(this._store.state.session.votingSpeed);
       this._send("nomination", nomination);
@@ -815,7 +816,13 @@ class LiveSession {
     const { session, players } = this._store.state;
     const playerCount = players.players.length;
     const indexAdjusted =
-      (index - 1 + playerCount - session.nomination[1]) % playerCount;
+      (index -
+        1 +
+        playerCount -
+        (typeof session.nomination[1] == "number"
+          ? session.nomination[1]
+          : session.nomination[0])) %
+      playerCount;
     if (fromST || indexAdjusted >= session.lockedVote - 1) {
       this._store.commit("session/vote", [index, vote]);
     }
@@ -828,7 +835,11 @@ class LiveSession {
     if (this._isSpectator) return;
     const { lockedVote, votes, nomination } = this._store.state.session;
     const { players } = this._store.state.players;
-    const index = (nomination[1] + lockedVote - 1) % players.length;
+    const index =
+      ((typeof nomination[1] == "number" ? nomination[1] : nomination[0]) +
+        lockedVote -
+        1) %
+      players.length;
     this._send("lock", [this._store.state.session.lockedVote, votes[index]]);
   }
 
@@ -844,7 +855,11 @@ class LiveSession {
     if (lock > 1) {
       const { lockedVote, nomination } = this._store.state.session;
       const { players } = this._store.state.players;
-      const index = (nomination[1] + lockedVote - 1) % players.length;
+      const index =
+        ((typeof nomination[1] == "number" ? nomination[1] : nomination[0]) +
+          lockedVote -
+          1) %
+        players.length;
       if (this._store.state.session.votes[index] !== vote) {
         this._store.commit("session/vote", [index, vote]);
       }